spark-apps/requirements.txt
# Spark Application Requirements
pyspark
delta-spark==3.5.2
